fixed issues of transformation not apply to selective port sending
i need to make a way to still apply transformation to a ports connected
list mainly for when sending to selected connected ports or socket,since
we use a subscriber based connection.
Solution was to clone the transformers and update this clones if any
change occurs on the origin transformers.
Fix applies to beginGroup,data and endGroup streams